Refactor: Improve code readability and add comments
