+++ src/components/consoles/views/MobileConsoleView.jsx
@@ -34,9 +34,3 @@ const MobileConsoleView = ({ stopClozer }) => {
 };
 
 export default MobileConsoleView;
-/*
-  <div className={`absolute top-0 left-0 w-8 h-8 bg-transparent border-t-2 border-l-2 border-600    transform -translate-x-0 -translate-y-0`}></div>
-      <div className={`absolute top-0 right-0 w-8 h-8 bg-transparent border-t-2 border-r-2 border-600   translate-x-0 -translate-y-0`}></div>
-      <div className={`absolute bottom-0 left-0 w-8 h-8 bg-transparent border-b-2 border-l-2 border-600  transform -translate-x-0 translate-y-0`}></div>
-      <div className={`absolute bottom-0 right-0 w-8 h-8 bg-transparent border-b-2 border-r-2 border-600  transform translate-x-0 translate-y-0`}></div>
-*/
