+++ src/main.jsx
@@ -4,7 +4,6 @@ import App from "./App.jsx";
 import { DataContextProvider } from "./context/DataContext.jsx";
 import { AudioContextProvider } from "./context/AudioContext.jsx";
 import { ConsoleContextProvider } from "./context/ConsoleContext.jsx";
-import 'typeface-montserrat'
 import "./index.css";
 
 ReactDOM.createRoot(document.getElementById("root")).render(
